+++ ft_printf_xcap.c
@@ -6,13 +6,13 @@
 /*   By: lpinheir <lpinheir@student.42sp.org.b      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/26 17:25:38 by lpinheir          #+#    #+#             */
-/*   Updated: 2021/05/01 11:21:13 by lpinheir         ###   ########.fr       */
+/*   Updated: 2021/05/01 11:56:09 by lpinheir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-static int	print_x(int precision, int strlen, long int num, char *hex)
+static int	print_x(int precision, int strlen, unsigned int num, char *hex)
 {
 	int	l_counter;
 	int	orig_precision;
@@ -27,7 +27,7 @@ static int	print_x(int precision, int strlen, long int num, char *hex)
 	return (l_counter);
 }
 
-static int	print_route(int inverted, int precision, int width, long int num)
+static int	print_route(int inverted, int precision, int width, unsigned int num)
 {
 	int		r_counter;
 	char	*hex;
@@ -56,7 +56,7 @@ int	ft_printf_xcap(int inverted, int width, int precision, va_list args)
 {
 	long int	num;
 
-	num = va_arg(args, int);
+	num = va_arg(args, unsigned int);
 	if (width < 0)
 	{
 		if (precision >= 0)
